+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### master
+- add notifications about the first and last match
 
 ### v0.0.2, Jun 26, 2014
 - search results are always at the bottom of the page. If possible center the